Fall back to task-repo when batch is absent

+++ harvester/tickle-repo/src/main/java/dk/dbc/dataio/harvester/ticklerepo/TaskRecordsIterator.java
@@ -67,7 +67,9 @@ public class TaskRecordsIterator implements RecordsIterator {
 
     @Override
     public void commit() {
-        taskRepo.getEntityManager().remove(task);
+        if (task != null) {
+            taskRepo.getEntityManager().remove(task);
+        }
     }
 
     public class ByList implements Iterator<Record> {
